+++ src/utils/BaseOutputStream.java
@@ -3,7 +3,7 @@ package utils;
 import java.io.IOException;
 
 public interface BaseOutputStream {
-    BaseOutputStream setBufferSize(int bufferSize);
+    BaseOutputStream setBufferSize(long bufferSize);
     void create(String filePath) throws IOException;
 
     void write(int value) throws IOException;
